refactor(collection): extract SearchManager
- Extract keyword, semantic, and hybrid search to SearchManager
- Delegate search, get_similar, and get_context to the manager
- Keep Collection facade clean while preserving existing API
